requirements.txt
asgiref==3.8.1
Django==5.1
django-address==0.2.8
django-colorfield==0.11.0
django-filter==24.3
django-leaflet==0.30.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
Markdown==3.7
phonenumbers==8.13.44
pillow==10.4.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
